refactor: Split the code into multiple files
Code is easier to read
